add authRoute + rename middleware
