+++ routes/auth-route.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const authRoute = express.Router()
+
+authRoute.get('*', ()=>{})
+
+
+module.exports = authRoute
